# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalysisError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "AnalysisError",
        "type": "object"
      },
      "AnalysisLinks": {
        "properties": {
          "events": {
            "title": "Events",
            "type": "string"
          },
          "export_json": {
            "title": "Export Json",
            "type": "string"
          },
          "export_markdown": {
            "title": "Export Markdown",
            "type": "string"
          },
          "export_sarif": {
            "title": "Export Sarif",
            "type": "string"
          },
          "self": {
            "title": "Self",
            "type": "string"
          }
        },
        "required": [
          "self",
          "events",
          "export_json",
          "export_markdown",
          "export_sarif"
        ],
        "title": "AnalysisLinks",
        "type": "object"
      },
...
      "AnalysisSummary": {
        "properties": {
          "ai_requested": {
            "default": false,
            "title": "Ai Requested",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisError"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "links": {
            "$ref": "#/components/schemas/AnalysisLinks"
          },
          "sample_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Id"
          },
          "source": {
            "enum": [
              "upload",
              "paste",
              "sample",
              "cli"
            ],
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HistorySummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at",
          "status",
          "title",
          "source",
          "sample_id",
          "ai_requested",
          "duration_ms",
          "summary",
          "error",
          "links"
        ],
        "title": "AnalysisSummary",
        "type": "object"
      },
...
      "HistorySummary": {
        "properties": {
          "additions": {
            "default": 0,
            "title": "Additions",
            "type": "integer"
          },
          "ai_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Status"
          },
          "by_kind": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Kind",
            "type": "object"
          },
          "by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "By Severity",
            "type": "object"
          },
          "deletions": {
            "default": 0,
            "title": "Deletions",
            "type": "integer"
          },
          "files_changed": {
            "default": 0,
            "title": "Files Changed",
            "type": "integer"
          },
          "findings_total": {
            "default": 0,
            "title": "Findings Total",
            "type": "integer"
          },
          "languages": {
            "items": {
              "type": "string"
            },
            "title": "Languages",
            "type": "array"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "patch_coverage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patch Coverage"
          },
          "review_priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Priority"
          }
        },
        "required": [
          "review_priority",
          "findings_total",
          "by_severity",
          "by_kind",
          "files_changed",
          "additions",
          "deletions",
          "languages",
          "mode",
          "ai_status",
          "patch_coverage"
        ],
        "title": "HistorySummary",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem details.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "request_id",
          "errors"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "SampleDetail": {
        "properties": {
          "coverage_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Filename"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "files": {
            "items": {
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          },
          "has_coverage": {
            "title": "Has Coverage",
            "type": "boolean"
          },
          "highlights": {
            "items": {
              "type": "string"
            },
            "title": "Highlights",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "patch": {
            "title": "Patch",
            "type": "string"
          },
          "synthetic": {
            "default": true,
            "title": "Synthetic",
            "type": "boolean"
          },
          "tagline": {
            "title": "Tagline",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "tagline",
          "description",
          "language",
          "tags",
          "highlights",
          "files",
          "has_coverage",
          "synthetic",
          "patch",
          "coverage_filename"
        ],
        "title": "SampleDetail",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Analyse a git diff (optionally with a repository snapshot and a coverage report) and receive a structured risk report. Every finding states its provenance (deterministic, heuristic, or AI-generated), severity and confidence separately, and cites evidence extracted from the inputs.\n\nAuthentication: when the server is configured with API keys, send `Authorization: Bearer <key>` or `X-API-Key`.",
    "summary": "Evidence-grounded code-change risk analysis.",
    "title": "ChangeGuard API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/samples/{sample_id}": {
      "get": {
        "operationId": "sample_detail_api_v1_samples__sample_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "sample_id",
            "required": true,
            "schema": {
              "title": "Sample Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Analysis not found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Input exceeds a size limit"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unsupported input format"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Malformed input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Get a sample scenario, including its patch",
        "tags": [
          "samples"
        ]
      }
    },
    "/api/v1/samples/{sample_id}/analyses": {
      "post": {
        "operationId": "analyse_sample_api_v1_samples__sample_id__analyses_post",
        "parameters": [
          {
            "in": "path",
            "name": "sample_id",
            "required": true,
            "schema": {
              "title": "Sample Id",
              "type": "string"
            }
          },
          {
            "description": "Run optional AI synthesis",
            "in": "query",
            "name": "ai",
            "required": false,
            "schema": {
              "default": false,
              "description": "Run optional AI synthesis",
              "title": "Ai",
              "type": "boolean"
            }
          },
          {
            "description": "Include the repository snapshot (full-context mode)",
            "in": "query",
            "name": "snapshot",
            "required": false,
            "schema": {
              "default": true,
              "description": "Include the repository snapshot (full-context mode)",
              "title": "Snapshot",
              "type": "boolean"
            }
          },
          {
            "description": "Include the sample's coverage report",
            "in": "query",
            "name": "coverage",
            "required": false,
            "schema": {
              "default": true,
              "description": "Include the sample's coverage report",
              "title": "Coverage",
              "type": "boolean"
            }
          },
          {
            "description": "Opaque workspace ID. Requests that carry one only see analyses created in it.",
            "in": "header",
            "name": "X-ChangeGuard-Workspace",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque workspace ID. Requests that carry one only see analyses created in it.",
              "title": "X-Changeguard-Workspace"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Analysis not found"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Input exceeds a size limit"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unsupported input format"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Malformed input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limit exceeded"
          }
        },
        "summary": "Analyse a sample scenario",
        "tags": [
          "samples"
        ]
      }
    }
  }
}